Afform - Move unit test to better place
Moves `testSubmitWithExtraFields` to `AfformExtraFieldInputTypeTest` from `AfformSubmitUsageTest`.
